continuing to fail
this fucking sucks. im still trying to get something to be recognized, but there's so much variety in formatting and there's typos and it's proving to be quite difficult. i have been making changes in an effort to create a useful csv. i am nearly there, i am only limited by the ability to pull the source and destination out of the route column.